fix/서비스 디렉토리 구조 변경
